+++ src/sage/symbolic/ginac/fderivative.cpp
@@ -250,13 +250,13 @@ bool fderivative::match_same_type(const basic & other) const
 	return parameter_set == o.parameter_set && inherited::match_same_type(other);
 }
 
-unsigned fderivative::calchash() const
+int64_t fderivative::calchash() const
 {
-	unsigned res = inherited::calchash();
-	unsigned prev=0;
+	int64_t res = inherited::calchash();
+	int64_t prev=0;
 
 	// FNV hash with custom prime and initial value
-	unsigned h = 2166136285;
+	int64_t h = 2166136285L;
 	for (const auto & elem : parameter_set) {
 		h = ( h * 2097287 ) ^ (elem-prev);
 		prev = elem;
