+++ src/sage/symbolic/ginac/fderivative.h
@@ -69,7 +69,7 @@ public:
 	ex thiscontainer(const exvector & v) const;
 	ex thiscontainer(std::unique_ptr<exvector> vp) const;
 	paramset get_parameter_set() const { return parameter_set; };
-	unsigned calchash() const;
+	int64_t calchash() const;
 
 protected:
 	ex derivative(const symbol & s) const;
