+++ src/sage/symbolic/ginac/idx.h
@@ -58,7 +58,7 @@ public:
 protected:
 	ex derivative(const symbol & s) const;
 	bool match_same_type(const basic & other) const;
-	unsigned calchash() const;
+	int64_t calchash() const;
 
 	// new virtual functions in this class
 public:
